requirements.txt
astor==0.8.1
astroid==2.9.3
astunparse==1.6.3
asynctest==0.13.0
black==22.1.0
blinker==1.4
brotlipy==0.7.0
certifi==2021.10.8
click==8.0.3
cloudpickle==1.6.0
cycler==0.11.0
flake8==4.0.1
fonttools==4.29.1
google-auth-oauthlib==0.4.1
gym==0.19.0
gym-notices==0.0.4
importlib-metadata==4.2.0
isort==5.10.1
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
matplotlib==3.5.1
mccabe==0.6.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mypy-extensions==0.4.3
numpy-stl==2.16.3
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.14.0
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyglet==1.5.22
pylint==2.12.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-utils==3.1.0
pytz==2021.3
requests-oauthlib==1.3.0
stable-baselines3==1.4.0
tensorboard-plugin-wit==1.6.0
tensorflow==1.14.0
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.63.0
typed-ast==1.5.2
stable-baselines3==1.4.0
gym==0.23.0
pyglet==1.5.22
# Editable install with no version control (multiagent==0.0.1)
-e ./multiagent-particle-envs